# Some performance improvements