@jgm 2022-10-15
Thanks - this is great.